Move NotificationUtils to bletia-core
