+++ bletia/src/main/java/info/izumin/android/bletia/action/EnableNotificationAction.java
@@ -5,7 +5,7 @@ import android.bluetooth.BluetoothGattDescriptor;
 
 import info.izumin.android.bletia.core.BleErrorType;
 import info.izumin.android.bletia.core.BletiaException;
-import info.izumin.android.bletia.util.NotificationUtils;
+import info.izumin.android.bletia.core.util.NotificationUtils;
 import info.izumin.android.bletia.core.wrapper.BluetoothGattWrapper;
 
 /**
